refactor(GH-15): :recycle: split loop review implementation files
Original ask: review the loop review changes, keep changed implementation files around the 300-line guideline, ensure there are no errors, stage the work, and update the existing PR.

Implementation summary:

- Split loop review, workflow loop, init refresh, capabilities catalog, and related tests into focused files under the changed-file size guideline.

- Kept generated init refresh defaults, capability metadata, PR title guidance, and loop review behavior covered by focused tests.

- Preserved the existing issue branch and PR lane for GH-15 / PR #16.

Verification performed:

- git diff --check

- git diff --staged --check

- secret scan over changed files

- /opt/homebrew/bin/go test ./...

- /opt/homebrew/bin/go run ./cmd/kit check 0035-loop-review

- /opt/homebrew/bin/go run ./cmd/kit map 0035-loop-review

- /opt/homebrew/bin/go run ./cmd/kit capabilities loop review --json

- /opt/homebrew/bin/go run ./cmd/kit capabilities --search review-loop --json

- /opt/homebrew/bin/go run ./cmd/kit loop --help

- /opt/homebrew/bin/go run ./cmd/kit loop review --help

- make build

Refs #15
