+++ docs/references/rules/github-pr-delivery.md
@@ -84,6 +84,8 @@ Delivery Contract:
 - If repo-local delivery rules cannot be found or are incomplete, stop and ask. Do not invent a substitute workflow.
 - Global agent/plugin GitHub workflows are fallback tools only. They do not define process in Kit-managed projects.
 - Do not create `codex/*` branches, ad hoc issue bodies, ad hoc PR bodies, draft PRs by default, commits using generic messages, or PRs that omit the repo template unless repo-local Kit rules explicitly require them or the user explicitly overrides the Kit contract.
+- The `PR title format` field must resolve to the Conventional Commits title shape with the GitHub issue as scope:
+  `<type>(<issue_number>): <gitmoji> <short title message>`.
 
 ### Author And Committer Invariant
 
@@ -272,7 +274,13 @@ git log -1 --format='%an <%ae> | %cn <%ce>'
 - Create the PR using `.github/pull_request_template.md` when present.
 - Preserve PR template headings.
 - Author the PR body in the user's name; do not add agent self-attribution.
-- Prefix the PR title with a descriptive gitmoji code that reflects the primary work.
+- The PR title must follow the Conventional Commits 1.0.0 title shape with the issue number as the required scope:
+  `<type>(<issue_number>): <gitmoji> <short title message>`.
+- Example: `feat(GH-123): :sparkles: add invitation workflow`.
+- Use the same allowed `type` values and deterministic type-to-gitmoji mapping as commit titles.
+- Keep the gitmoji immediately after the colon and space as part of the Conventional Commit description.
+- Do not create gitmoji-only PR titles, sentence-case PR titles, or PR titles without the `GH-123` issue scope.
+- Use `!` for a breaking-change PR title only when the change is intentionally breaking and the user explicitly approves that signal.
 - Prefix each concrete bullet in the PR `Description` section with a descriptive gitmoji code that reflects that bullet's work.
 - Use gitmoji codes that are aligned with the work being described, such as `:sparkles:` for feature work, `:bug:` for fixes, `:memo:` for documentation, `:white_check_mark:` for tests, `:recycle:` for refactors, `:green_heart:` for CI, `:wrench:` for tooling, `:dizzy:` for refresh or migration flows, and `:notebook:` for constitution, guide, or reference updates.
 - Keep PR headings from the template unchanged; add gitmoji prefixes to title and description content, not to template headings.
@@ -352,7 +360,7 @@ Include:
 - Confirm branch was pushed only after commit.
 - Confirm active directory, branch, remote, and PR state were rechecked immediately before pushing.
 - Confirm PR was created with the repository template when present.
-- Confirm PR title and `Description` bullets use descriptive gitmoji prefixes while preserving template headings.
+- Confirm PR title follows the required Conventional Commit format and `Description` bullets use descriptive gitmoji prefixes while preserving template headings.
 - Confirm PR assignee is the human user.
 - Confirm PR is ready for review and not draft unless explicitly requested otherwise.
 - Confirm PR and CI state were observed before final reporting.
@@ -408,7 +416,7 @@ docs(GH-125): :memo: document PR workflow
 PR title and description formatting:
 
 ```text
-PR Title: :sparkles: add invitation workflow
+PR Title: feat(GH-123): :sparkles: add invitation workflow
 
 ## Description
