+++ docs/specs/0035-loop-review/PLAN.md
@@ -35,14 +35,18 @@ Refactor `kit loop` into a command group, keep the existing workflow runner as `
 - Reuse `.kit/loops/<run-id>/` artifact helpers.
 - Reuse existing PR target parsing, CodeRabbit review-task extraction, and watch timing helpers.
 - Keep default PR mode opportunistic; only `--watch` / `--wait-for-coderabbit` blocks for CodeRabbit completion.
+- Route review prompts through shared prompt helpers without subagents by default; add opt-in `--subagents` pre-analysis before the review final-output contract.
+- Treat agent command setup failures as terminal setup failures, not as review passes.
+- Stream runner progress and child-agent stdout/stderr to stderr for human-readable runs while preserving quiet JSON output.
 
 ## COMPONENTS
 
 - [PLAN-COMPONENTS-01] `pkg/cli/loop.go`: command group wiring, workflow subcommand, default iteration change.
-- [PLAN-COMPONENTS-02] `pkg/cli/loop_review.go`: review options, diff discovery, prompt builder, agent loop, result parser, artifact writer, and PR feedback polling.
+- [PLAN-COMPONENTS-02] `pkg/cli/loop_review.go`: review options, diff discovery, prompt builder, opt-in subagent prompt routing, progress streaming, agent loop, rerun prompts, result parser, artifact writer, and PR feedback polling.
 - [PLAN-COMPONENTS-03] `pkg/cli/capabilities_catalog.go`: loop and loop review metadata.
 - [PLAN-COMPONENTS-04] Docs: README, agent tooling, and init spec command reference.
 - [PLAN-COMPONENTS-05] Tests: routing, parsing, base fallback, PR feedback handling, and capabilities.
+- [PLAN-COMPONENTS-06] `pkg/cli/init.go` and `pkg/cli/init_refresh.go`: seed and refresh default loop agent configuration.
 
 ## DATA
 
@@ -54,6 +58,7 @@ Refactor `kit loop` into a command group, keep the existing workflow runner as `
 - Builds on the existing `0034-review-loop` PR review-thread intake and CodeRabbit filtering helpers.
 - Relies on the configured loop agent in `.kit.yaml`.
 - Uses Git for local changed-code discovery and `gh` only when PR mode is requested.
+- Existing projects can adopt the default Codex loop agent config through `kit init --refresh`.
 
 ## INTERFACES
 
@@ -76,4 +81,5 @@ Refactor `kit loop` into a command group, keep the existing workflow runner as `
 - Unit tests for base fallback with fake command runner.
 - Unit tests for opportunistic PR feedback ingestion using existing fake review-loop seams.
 - Capabilities tests for new nested command metadata.
+- Init refresh tests for loop agent config backfill and custom-agent preservation.
 - Full `go test ./...` verification.
