+++ pkg/cli/capabilities_catalog.go
@@ -2,224 +2,13 @@ package cli
 
 import (
 	"sort"
-	"strings"
 )
 
 const capabilitiesSchemaVersion = 1
 
-const (
-	mutationNone             = "none"
-	mutationWritesFiles      = "writes_files"
-	mutationExecutesCommands = "executes_commands"
-	mutationNetwork          = "network"
-	mutationGit              = "git"
-	mutationDestructive      = "destructive"
-)
-
-var capabilityCategoryOrder = map[string]int{
-	"Setup":            1,
-	"Workflow":         2,
-	"Inspect & Repair": 3,
-	"Prompt Utilities": 4,
-	"Utilities":        5,
-}
-
-type capabilityBehavior struct {
-	Summary       string `json:"summary"`
-	FlagDependent string `json:"flag_dependent,omitempty"`
-}
-
-type capabilityFlag struct {
-	Name    string `json:"name"`
-	Summary string `json:"summary"`
-	Safety  string `json:"safety,omitempty"`
-}
-
-type capabilityRelatedCommand struct {
-	Command string `json:"command"`
-	Note    string `json:"note,omitempty"`
-}
-
-type capabilityRecord struct {
-	Command              string
-	Category             string
-	Summary              string
-	MutationLevel        string
-	NetworkUse           capabilityBehavior
-	FileWrites           capabilityBehavior
-	GitMutation          capabilityBehavior
-	Hidden               bool
-	Deprecated           bool
-	DeprecationNote      string
-	Aliases              []string
-	ImportantFlags       []capabilityFlag
-	RelatedCommands      []capabilityRelatedCommand
-	WhenToUse            []string
-	WhenNotToUse         []string
-	Examples             []string
-	Caveats              []string
-	DetailedFlagBehavior []capabilityFlag
-	IncludeInCompact     bool
-}
-
-type capabilityCompactRecord struct {
-	Command         string                     `json:"command"`
-	Category        string                     `json:"category"`
-	Summary         string                     `json:"summary"`
-	MutationLevel   string                     `json:"mutation_level"`
-	NetworkUse      capabilityBehavior         `json:"network_use"`
-	FileWrites      capabilityBehavior         `json:"file_writes"`
-	GitMutation     capabilityBehavior         `json:"git_mutation"`
-	Hidden          bool                       `json:"hidden"`
-	Deprecated      bool                       `json:"deprecated"`
-	ImportantFlags  []capabilityFlag           `json:"important_flags"`
-	RelatedCommands []capabilityRelatedCommand `json:"related_commands"`
-}
-
-type capabilityDetailRecord struct {
-	Command              string                     `json:"command"`
-	Category             string                     `json:"category"`
-	Summary              string                     `json:"summary"`
-	MutationLevel        string                     `json:"mutation_level"`
-	NetworkUse           capabilityBehavior         `json:"network_use"`
-	FileWrites           capabilityBehavior         `json:"file_writes"`
-	GitMutation          capabilityBehavior         `json:"git_mutation"`
-	Hidden               bool                       `json:"hidden"`
-	Deprecated           bool                       `json:"deprecated"`
-	DeprecationNote      string                     `json:"deprecation_note,omitempty"`
-	Aliases              []string                   `json:"aliases"`
-	ImportantFlags       []capabilityFlag           `json:"important_flags"`
-	RelatedCommands      []capabilityRelatedCommand `json:"related_commands"`
-	WhenToUse            []string                   `json:"when_to_use"`
-	WhenNotToUse         []string                   `json:"when_not_to_use"`
-	Examples             []string                   `json:"examples"`
-	Caveats              []string                   `json:"caveats"`
-	DetailedFlagBehavior []capabilityFlag           `json:"detailed_flag_behavior"`
-}
-
-type capabilityOption func(*capabilityRecord)
-
-func capability(command, category, summary, mutationLevel string, options ...capabilityOption) capabilityRecord {
-	record := capabilityRecord{
-		Command:          command,
-		Category:         category,
-		Summary:          summary,
-		MutationLevel:    mutationLevel,
-		NetworkUse:       capabilityBehavior{Summary: "none"},
-		FileWrites:       capabilityBehavior{Summary: "none"},
-		GitMutation:      capabilityBehavior{Summary: "none"},
-		Aliases:          []string{},
-		ImportantFlags:   []capabilityFlag{},
-		RelatedCommands:  []capabilityRelatedCommand{},
-		WhenToUse:        []string{summary},
-		WhenNotToUse:     []string{"Use a narrower Kit command when one better matches the workflow step."},
-		Examples:         []string{"kit " + command},
-		Caveats:          []string{},
-		IncludeInCompact: true,
-	}
-	for _, option := range options {
-		option(&record)
-	}
-	if record.DetailedFlagBehavior == nil {
-		record.DetailedFlagBehavior = append([]capabilityFlag(nil), record.ImportantFlags...)
-	}
-	return record
-}
-
-func withNetwork(summary string, flagDependent ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.NetworkUse = capabilityBehavior{Summary: summary}
-		if len(flagDependent) > 0 {
-			record.NetworkUse.FlagDependent = flagDependent[0]
-		}
-	}
-}
-
-func withFileWrites(summary string, flagDependent ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.FileWrites = capabilityBehavior{Summary: summary}
-		if len(flagDependent) > 0 {
-			record.FileWrites.FlagDependent = flagDependent[0]
-		}
-	}
-}
-
-func withGitMutation(summary string, flagDependent ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.GitMutation = capabilityBehavior{Summary: summary}
-		if len(flagDependent) > 0 {
-			record.GitMutation.FlagDependent = flagDependent[0]
-		}
-	}
-}
-
-func withFlags(flags ...capabilityFlag) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.ImportantFlags = append(record.ImportantFlags, flags...)
-		record.DetailedFlagBehavior = append(record.DetailedFlagBehavior, flags...)
-	}
-}
-
-func withRelated(related ...capabilityRelatedCommand) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.RelatedCommands = append(record.RelatedCommands, related...)
-	}
-}
-
-func withAliases(aliases ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.Aliases = append(record.Aliases, aliases...)
-	}
-}
-
-func withWhenToUse(values ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.WhenToUse = append([]string(nil), values...)
-	}
-}
-
-func withWhenNotToUse(values ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.WhenNotToUse = append([]string(nil), values...)
-	}
-}
-
-func withExamples(values ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.Examples = append([]string(nil), values...)
-	}
-}
-
-func withCaveats(values ...string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.Caveats = append(record.Caveats, values...)
-	}
-}
-
-func hiddenDeprecated(note string) capabilityOption {
-	return func(record *capabilityRecord) {
-		record.Hidden = true
-		record.Deprecated = true
-		record.DeprecationNote = note
-		record.IncludeInCompact = false
-	}
-}
-
-func flag(name, summary string, safety ...string) capabilityFlag {
-	result := capabilityFlag{Name: name, Summary: summary}
-	if len(safety) > 0 {
-		result.Safety = safety[0]
-	}
-	return result
-}
-
-func related(command, note string) capabilityRelatedCommand {
-	return capabilityRelatedCommand{Command: command, Note: note}
-}
-
 func capabilityCatalog() []capabilityRecord {
 	records := []capabilityRecord{
-		capability("init", "Setup", "Initialize Kit project scaffolding in a repository.", mutationWritesFiles, withFileWrites("writes Kit project files and docs"), withFlags(flag("--force", "replace existing generated files when supported", "review local edits first")), withRelated(related("scaffold", "generates individual workflow artifacts"))),
+		capability("init", "Setup", "Initialize or refresh Kit project scaffolding in a repository.", mutationWritesFiles, withNetwork("none by default", "--refresh can fetch the Kit ruleset registry from GitHub"), withFileWrites("writes .kit.yaml, Kit project files, docs, instruction files, and registry rulesets", "--dry-run previews refresh changes without writing files"), withFlags(flag("--refresh", "refresh Kit-managed files and backfill or upgrade generated .kit.yaml settings such as loop.agent.command"), flag("--dry-run", "preview --refresh without writing files", "read-only"), flag("--diff", "print planned --refresh changes as a unified diff; requires --dry-run", "read-only"), flag("--file", "limit --refresh to one Kit-managed file; repeat for multiple files"), flag("--force", "replace existing generated files when supported", "review local edits first")), withRelated(related("scaffold", "generates individual workflow artifacts"), related("loop review", "uses the loop agent config seeded by init refresh")), withWhenToUse("Use once to initialize Kit scaffolding in a repository.", "Use `kit init --refresh` in an existing Kit project to adopt current managed files, registry rulesets, and default loop agent configuration."), withWhenNotToUse("Do not use for semantic project documentation refresh after the repo matures; use `kit prompt project refresh`.", "Do not use `--force` until local generated-doc changes have been reviewed."), withExamples("kit init", "kit init --refresh", "kit init --refresh --dry-run --diff", "kit init --refresh --file=.kit.yaml")),
 		capability("scaffold", "Setup", "Generate workflow artifacts for Kit features.", mutationWritesFiles, withFileWrites("writes generated docs under selected project paths"), withFlags(flag("--feature", "select the feature slug when supported")), withRelated(related("scaffold agents", "writes repo agent instructions"))),
 		capability("scaffold agents", "Setup", "Generate or refresh repo-local agent instruction files.", mutationWritesFiles, withFileWrites("writes AGENTS.md and docs/agents guidance"), withFlags(flag("--force", "replace existing agent guidance files", "review local edits first")), withRelated(related("init", "creates the broader project structure"))),
 		capability("brainstorm", "Workflow", "Create or inspect the brainstorm artifact for a feature.", mutationWritesFiles, withFileWrites("writes docs/specs/<feature>/BRAINSTORM.md when generating"), withRelated(related("spec", "turns brainstorm material into requirements"))),
@@ -229,7 +18,38 @@ func capabilityCatalog() []capabilityRecord {
 		capability("tasks", "Workflow", "Create or update the task list for a feature.", mutationWritesFiles, withFileWrites("writes docs/specs/<feature>/TASKS.md"), withRelated(related("implement", "uses tasks as the execution checklist"))),
 		capability("loop", "Workflow", "Run workflow and review agent loops.", mutationExecutesCommands, withFileWrites("can write loop state under .kit/loops"), withGitMutation("none", "delegated agent commands may mutate git independently"), withFlags(flag("--dry-run", "show legacy workflow loop plan without delegated execution", "read-only"), flag("--max-iterations", "bound legacy workflow loop execution")), withRelated(related("loop workflow", "canonical feature workflow loop"), related("loop review", "changed-code correctness review loop"), related("implement", "runs outside the loop runner")), withExamples("kit loop workflow my-feature", "kit loop review", "kit loop review --pr 14")),
 		capability("loop workflow", "Workflow", "Run the feature workflow through a confidence-gated local agent loop.", mutationExecutesCommands, withFileWrites("writes loop prompts, stdout, stderr, and run summaries under .kit/loops"), withGitMutation("none", "delegated agent commands may mutate files or git independently; Kit itself does not stage, commit, or push"), withFlags(flag("--dry-run", "show the next workflow stage without invoking the configured agent", "read-only"), flag("--until", "stop after a workflow stage is complete"), flag("--min-confidence", "override the required agent confidence percentage"), flag("--max-iterations", "bound loop execution"), flag("--json", "emit the loop report as JSON")), withRelated(related("loop review", "runs correctness review passes over changed code"), related("verify", "records validation evidence")), withExamples("kit loop workflow my-feature --dry-run", "kit loop workflow my-feature --until tasks")),
-		capability("loop review", "Workflow", "Run a coding-agent correctness review loop over changes not in the remote mainline.", mutationExecutesCommands, withNetwork("none by default", "--pr reads PR metadata, CodeRabbit checks, and unresolved non-outdated CodeRabbit review threads through gh"), withFileWrites("writes loop prompts, stdout, stderr, and run summaries under .kit/loops; delegated agent may edit local project files"), withGitMutation("none; command prompts explicitly forbid staging, commits, pushes, PR comments, and review-thread resolution"), withFlags(flag("--base", "override the comparison base; default is origin/main, then main"), flag("--pr", "opportunistically ingest CodeRabbit feedback from a pull request", "network read"), flag("--watch", "wait for CodeRabbit completion before finalizing PR-mode review", "network read with polling"), flag("--wait-for-coderabbit", "alias for --watch", "network read with polling"), flag("--dry-run", "show the first review prompt without invoking the configured agent", "read-only"), flag("--min-confidence", "override the required correctness percentage"), flag("--max-iterations", "bound review passes; default is 10"), flag("--json", "emit the loop review report as JSON")), withRelated(related("loop workflow", "runs the feature artifact workflow loop"), related("review-loop", "legacy PR prompt-prep compatibility command"), related("dispatch --loop", "PR prompt-prep alias")), withWhenToUse("Use before publishing or updating a PR when changed code should be locally reviewed and repaired until no high, medium, or correctness-impacting issues remain.", "Use with --pr when CodeRabbit feedback should be folded into local repair passes opportunistically."), withWhenNotToUse("Do not use when you only need a prompt from PR review threads; use `dispatch --loop --pr` or legacy `review-loop`.", "Do not use as a git/GitHub delivery command; it never stages, commits, pushes, comments, or resolves threads."), withExamples("kit loop review", "kit loop review --base origin/main", "kit loop review --pr 14", "kit loop review --pr 14 --wait-for-coderabbit"), withCaveats("The correctness percentage is agent-reported and gated by the prompt plus validation evidence; it is not a mathematical proof.")),
+		capability(
+			"loop review",
+			"Workflow",
+			"Run a coding-agent correctness review loop over changes not in the remote mainline.",
+			mutationExecutesCommands,
+			withNetwork("none by default", "--pr reads PR metadata, CodeRabbit checks, and unresolved non-outdated CodeRabbit review threads through gh"),
+			withFileWrites("writes loop prompts, stdout, stderr, and run summaries under .kit/loops; delegated agent may edit local project files"),
+			withGitMutation("none; command prompts explicitly forbid staging, commits, pushes, PR comments, and review-thread resolution"),
+			withFlags(
+				flag("--base", "override the comparison base; default is origin/main, then main"),
+				flag("--pr", "opportunistically ingest CodeRabbit feedback from a pull request", "network read"),
+				flag("--watch", "wait for CodeRabbit completion before finalizing PR-mode review", "network read with polling"),
+				flag("--wait-for-coderabbit", "alias for --watch", "network read with polling"),
+				flag("--dry-run", "show the first review prompt without invoking the configured agent", "read-only"),
+				flag("--min-confidence", "override the required correctness percentage"),
+				flag("--max-iterations", "bound review passes; default is 10"),
+				flag("--subagents", "allow the parent review agent to pre-analyze changed-code lanes and choose subagents"),
+				flag("--json", "emit the loop review report as JSON"),
+			),
+			withRelated(related("loop workflow", "runs the feature artifact workflow loop"), related("review-loop", "legacy PR prompt-prep compatibility command"), related("dispatch --loop", "PR prompt-prep alias")),
+			withWhenToUse("Use before publishing or updating a PR when changed code should be locally reviewed and repaired until no high, medium, or correctness-impacting issues remain.", "Use with --pr when CodeRabbit feedback should be folded into local repair passes opportunistically."),
+			withWhenNotToUse("Do not use when you only need a prompt from PR review threads; use `dispatch --loop --pr` or legacy `review-loop`.", "Do not use as a git/GitHub delivery command; it never stages, commits, pushes, comments, or resolves threads."),
+			withExamples("kit loop review", "kit loop review --base origin/main", "kit loop review --pr 14", "kit loop review --pr 14 --wait-for-coderabbit"),
+			withCaveats(
+				"The correctness percentage is agent-reported and gated by the prompt plus validation evidence; it is not a mathematical proof.",
+				"Review prompts use one agent by default; --subagents adds parent pre-analysis and shared subagent guidance when review lanes are clearly independent.",
+				"The generated default Codex loop agent pins --model gpt-5.5 and uses --ignore-user-config to avoid unsupported model defaults and unrelated local MCP/plugin auth side effects during child-agent runs.",
+				"Human-readable runs stream emoji-marked runner progress and child-agent stdout/stderr to stderr; --json suppresses progress output so stdout stays machine-readable.",
+				"Interactive terminals are prompted before rerunning after existing loop evidence or a max-iterations stop; JSON and dry-run modes do not prompt.",
+				"Agent command setup failures stop immediately with stderr context instead of consuming every iteration.",
+			),
+		),
 		capability("resume", "Workflow", "Resume the next useful Kit workflow action.", mutationNone, withRelated(related("status", "shows current feature state before resuming"))),
 		capability("implement", "Workflow", "Generate or print the implementation prompt for a feature.", mutationNone, withRelated(related("tasks", "defines the implementation checklist"), related("verify", "validates completed work"))),
 		capability("reflect", "Workflow", "Generate or print a reflection prompt for completed feature work.", mutationNone, withRelated(related("complete", "marks finished work after validation"))),
@@ -278,225 +98,4 @@ func capabilityCatalog() []capabilityRecord {
 		return lessCapabilityRecord(records[i], records[j])
 	})
 	return records
-}
-
-func compactCapabilityRecords(records []capabilityRecord) []capabilityCompactRecord {
-	compact := make([]capabilityCompactRecord, 0, len(records))
-	for _, record := range records {
-		if !record.IncludeInCompact || record.Hidden || record.Deprecated {
-			continue
-		}
-		compact = append(compact, record.compact())
-	}
-	return compact
-}
-
-func detailCapabilityRecords(records []capabilityRecord) []capabilityDetailRecord {
-	detail := make([]capabilityDetailRecord, 0, len(records))
-	for _, record := range records {
-		detail = append(detail, record.detail())
-	}
-	return detail
-}
-
-func capabilityByCommandPath(commandPath string) (capabilityRecord, bool) {
-	normalized := normalizeCapabilityQuery(commandPath)
-	for _, record := range capabilityCatalog() {
-		if record.matchesCommandPath(normalized) {
-			return record, true
-		}
-	}
-	return capabilityRecord{}, false
-}
-
-func searchCapabilityRecords(query string) []capabilityRecord {
-	normalized := normalizeCapabilityQuery(query)
-	if normalized == "" {
-		return visibleCapabilityRecords()
-	}
-
-	var matches []capabilityRecord
-	for _, record := range visibleCapabilityRecords() {
-		if record.matchesSearch(normalized) {
-			matches = append(matches, record)
-		}
-	}
-	return matches
-}
-
-func visibleCapabilityRecords() []capabilityRecord {
-	records := capabilityCatalog()
-	visible := make([]capabilityRecord, 0, len(records))
-	for _, record := range records {
-		if record.IncludeInCompact && !record.Hidden && !record.Deprecated {
-			visible = append(visible, record)
-		}
-	}
-	return visible
-}
-
-func suggestCapabilityCommands(commandPath string) []string {
-	normalized := normalizeCapabilityQuery(commandPath)
-	if normalized == "" {
-		return nil
-	}
-
-	suggestions := make([]string, 0, 3)
-	for _, record := range capabilityCatalog() {
-		candidates := append([]string{record.Command}, record.Aliases...)
-		for _, candidate := range candidates {
-			candidate = normalizeCapabilityQuery(candidate)
-			if strings.HasPrefix(candidate, normalized) || strings.HasPrefix(normalized, candidate) || strings.Contains(candidate, normalized) {
-				suggestions = appendUniqueSuggestion(suggestions, record.Command)
-			}
-			if len(suggestions) >= 3 {
-				return suggestions
-			}
-		}
-	}
-	return suggestions
-}
-
-func appendUniqueSuggestion(suggestions []string, command string) []string {
-	for _, existing := range suggestions {
-		if existing == command {
-			return suggestions
-		}
-	}
-	return append(suggestions, command)
-}
-
-func (record capabilityRecord) compact() capabilityCompactRecord {
-	return capabilityCompactRecord{
-		Command:         record.Command,
-		Category:        record.Category,
-		Summary:         record.Summary,
-		MutationLevel:   record.MutationLevel,
-		NetworkUse:      record.NetworkUse,
-		FileWrites:      record.FileWrites,
-		GitMutation:     record.GitMutation,
-		Hidden:          record.Hidden,
-		Deprecated:      record.Deprecated,
-		ImportantFlags:  cloneCapabilityFlags(record.ImportantFlags),
-		RelatedCommands: cloneRelatedCommands(record.RelatedCommands),
-	}
-}
-
-func (record capabilityRecord) detail() capabilityDetailRecord {
-	return capabilityDetailRecord{
-		Command:              record.Command,
-		Category:             record.Category,
-		Summary:              record.Summary,
-		MutationLevel:        record.MutationLevel,
-		NetworkUse:           record.NetworkUse,
-		FileWrites:           record.FileWrites,
-		GitMutation:          record.GitMutation,
-		Hidden:               record.Hidden,
-		Deprecated:           record.Deprecated,
-		DeprecationNote:      record.DeprecationNote,
-		Aliases:              cloneStrings(record.Aliases),
-		ImportantFlags:       cloneCapabilityFlags(record.ImportantFlags),
-		RelatedCommands:      cloneRelatedCommands(record.RelatedCommands),
-		WhenToUse:            cloneStrings(record.WhenToUse),
-		WhenNotToUse:         cloneStrings(record.WhenNotToUse),
-		Examples:             cloneStrings(record.Examples),
-		Caveats:              cloneStrings(record.Caveats),
-		DetailedFlagBehavior: cloneCapabilityFlags(record.DetailedFlagBehavior),
-	}
-}
-
-func lessCapabilityRecord(left, right capabilityRecord) bool {
-	leftCategory := capabilityCategoryOrder[left.Category]
-	rightCategory := capabilityCategoryOrder[right.Category]
-	if leftCategory != rightCategory {
-		return leftCategory < rightCategory
-	}
-
-	leftOrder := capabilityCommandOrder(left.Command)
-	rightOrder := capabilityCommandOrder(right.Command)
-	if leftOrder != rightOrder {
-		return leftOrder < rightOrder
-	}
-	return left.Command < right.Command
-}
-
-func capabilityCommandOrder(commandPath string) int {
-	rootName := commandPath
-	if fields := strings.Fields(commandPath); len(fields) > 0 {
-		rootName = fields[0]
-	}
-	if order, ok := commandOrder[rootName]; ok {
-		return order
-	}
-	return 1000
-}
-
-func cloneStrings(values []string) []string {
-	if len(values) == 0 {
-		return []string{}
-	}
-	return append([]string(nil), values...)
-}
-
-func cloneCapabilityFlags(values []capabilityFlag) []capabilityFlag {
-	if len(values) == 0 {
-		return []capabilityFlag{}
-	}
-	return append([]capabilityFlag(nil), values...)
-}
-
-func cloneRelatedCommands(values []capabilityRelatedCommand) []capabilityRelatedCommand {
-	if len(values) == 0 {
-		return []capabilityRelatedCommand{}
-	}
-	return append([]capabilityRelatedCommand(nil), values...)
-}
-
-func (record capabilityRecord) matchesCommandPath(commandPath string) bool {
-	if normalizeCapabilityQuery(record.Command) == commandPath {
-		return true
-	}
-	for _, alias := range record.Aliases {
-		if normalizeCapabilityQuery(alias) == commandPath {
-			return true
-		}
-	}
-	return false
-}
-
-func (record capabilityRecord) matchesSearch(query string) bool {
-	searchable := []string{
-		record.Command,
-		record.Category,
-		record.Summary,
-		record.MutationLevel,
-		record.NetworkUse.Summary,
-		record.NetworkUse.FlagDependent,
-		record.FileWrites.Summary,
-		record.FileWrites.FlagDependent,
-		record.GitMutation.Summary,
-		record.GitMutation.FlagDependent,
-	}
-	searchable = append(searchable, record.Aliases...)
-	searchable = append(searchable, record.WhenToUse...)
-	searchable = append(searchable, record.WhenNotToUse...)
-	searchable = append(searchable, record.Examples...)
-	searchable = append(searchable, record.Caveats...)
-	for _, flag := range record.ImportantFlags {
-		searchable = append(searchable, flag.Name, flag.Summary, flag.Safety)
-	}
-	for _, related := range record.RelatedCommands {
-		searchable = append(searchable, related.Command, related.Note)
-	}
-
-	for _, value := range searchable {
-		if strings.Contains(normalizeCapabilityQuery(value), query) {
-			return true
-		}
-	}
-	return false
-}
-
-func normalizeCapabilityQuery(value string) string {
-	return strings.Join(strings.Fields(strings.ToLower(strings.TrimSpace(value))), " ")
 }
