+++ pkg/cli/init_refresh.go
@@ -3,13 +3,11 @@ package cli
 import (
 	"context"
 	"fmt"
-	"os"
 	"path/filepath"
 	"sort"
 	"strings"
 
 	"github.com/jamesonstone/kit/internal/config"
-	"gopkg.in/yaml.v3"
 )
 
 const constitutionBaselineHeading = "Kit-Managed Baseline Rules"
@@ -144,67 +142,6 @@ func runInitRefresh(projectRoot string, opts initRefreshOptions) error {
 	return nil
 }
 
-func initRefreshConfig(
-	projectRoot string,
-	opts initRefreshOptions,
-	targets map[string]bool,
-) (*config.Config, *initRefreshFileChange, error) {
-	cfg := defaultInitConfig()
-	configSelected := initRefreshTargetMatches(targets, config.ConfigFileName)
-	shouldTouchConfig := len(targets) == 0 || configSelected
-	path := filepath.Join(projectRoot, config.ConfigFileName)
-	exists := config.Exists(projectRoot)
-	var before string
-
-	if exists {
-		data, err := os.ReadFile(path)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to read %s: %w", config.ConfigFileName, err)
-		}
-		before = string(data)
-		existing, err := config.Load(projectRoot)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to load %s: %w", config.ConfigFileName, err)
-		}
-		cfg = existing
-	}
-
-	if configSelected && opts.force {
-		cfg = defaultInitConfig()
-		cfg.InstructionScaffoldVersion = config.InstructionScaffoldVersionTOC
-		after, err := marshalInitRefreshConfig(cfg)
-		if err != nil {
-			return nil, nil, err
-		}
-		result := instructionFileCreated
-		if exists {
-			result = instructionFileUpdated
-		}
-		return cfg, newInitRefreshFileChange(projectRoot, config.ConfigFileName, before, after, result), nil
-	}
-
-	if cfg.InstructionScaffoldVersion != config.InstructionScaffoldVersionTOC {
-		cfg.InstructionScaffoldVersion = config.InstructionScaffoldVersionTOC
-		if shouldTouchConfig {
-			after, err := marshalInitRefreshConfig(cfg)
-			if err != nil {
-				return nil, nil, err
-			}
-			return cfg, newInitRefreshFileChange(projectRoot, config.ConfigFileName, before, after, instructionFileUpdated), nil
-		}
-	}
-
-	if !exists && shouldTouchConfig {
-		after, err := marshalInitRefreshConfig(cfg)
-		if err != nil {
-			return nil, nil, err
-		}
-		return cfg, newInitRefreshFileChange(projectRoot, config.ConfigFileName, before, after, instructionFileCreated), nil
-	}
-
-	return cfg, nil, nil
-}
-
 func initRefreshKnownTargets(cfg *config.Config, registry []registryRuleset) map[string]bool {
 	known := map[string]bool{
 		config.ConfigFileName:                  true,
@@ -267,31 +204,6 @@ func validateInitRefreshTargets(targets, known map[string]bool) error {
 	return fmt.Errorf("%s is not a Kit-managed refresh target", strings.Join(unknown, ", "))
 }
 
-func finalizeInitRefreshConfigChange(projectRoot string, cfg *config.Config, planned *initRefreshFileChange) (*initRefreshFileChange, error) {
-	before := ""
-	result := instructionFileCreated
-	if planned != nil {
-		before = planned.before
-		result = planned.result
-	} else if config.Exists(projectRoot) {
-		data, err := os.ReadFile(filepath.Join(projectRoot, config.ConfigFileName))
-		if err != nil {
-			return nil, fmt.Errorf("failed to read %s: %w", config.ConfigFileName, err)
-		}
-		before = string(data)
-		result = instructionFileUpdated
-	}
-
-	after, err := marshalInitRefreshConfig(cfg)
-	if err != nil {
-		return nil, err
-	}
-	if before == after {
-		return nil, nil
-	}
-	return newInitRefreshFileChange(projectRoot, config.ConfigFileName, before, after, result), nil
-}
-
 func printInitRefreshNotes(notes []string, opts initRefreshOptions) {
 	if opts.outputOnly || len(notes) == 0 {
 		return
@@ -326,12 +238,4 @@ func (s *initRefreshStats) recordResult(result instructionFileWriteResult) {
 	case instructionFileSkipped:
 		s.skipped++
 	}
-}
-
-func marshalInitRefreshConfig(cfg *config.Config) (string, error) {
-	data, err := yaml.Marshal(cfg)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal config: %w", err)
-	}
-	return string(data), nil
 }
