+++ pkg/cli/init_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/jamesonstone/kit/internal/config"
-	"github.com/jamesonstone/kit/internal/document"
 	"github.com/jamesonstone/kit/internal/templates"
 )
 
@@ -168,7 +167,7 @@ func TestRunInit_PopulatesGlobalConfig(t *testing.T) {
 	})
 }
 
-func TestRunInit_CreatesCodeRabbitConfig(t *testing.T) {
+func TestRunInit_CreatesLoopReviewAgentConfig(t *testing.T) {
 	tempDir := t.TempDir()
 	setupInitHome(t)
 	setWorkingDirectory(t, tempDir)
@@ -181,238 +180,13 @@ func TestRunInit_CreatesCodeRabbitConfig(t *testing.T) {
 				t.Fatalf("runInit() error = %v", err)
 			}
 		})
-
-		content, err := os.ReadFile(filepath.Join(tempDir, codeRabbitConfigPath))
-		if err != nil {
-			t.Fatalf("expected %s to be created: %v", codeRabbitConfigPath, err)
-		}
-		if string(content) != templates.CodeRabbitConfig {
-			t.Fatalf("%s content = %q, want %q", codeRabbitConfigPath, content, templates.CodeRabbitConfig)
-		}
 	})
-}
 
-func TestRunInit_CreatesPullRequestTemplate(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		content, err := os.ReadFile(filepath.Join(tempDir, pullRequestTemplatePath))
-		if err != nil {
-			t.Fatalf("expected %s to be created: %v", pullRequestTemplatePath, err)
-		}
-		if string(content) != templates.PullRequestTemplate {
-			t.Fatalf("%s content = %q, want %q", pullRequestTemplatePath, content, templates.PullRequestTemplate)
-		}
-	})
-}
-
-func TestRunInit_CreatesGitignoreWithKitLocalArtifacts(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		content, err := os.ReadFile(filepath.Join(tempDir, gitignorePath))
-		if err != nil {
-			t.Fatalf("expected %s to be created: %v", gitignorePath, err)
-		}
-		got := string(content)
-		for _, pattern := range kitGitignorePatterns() {
-			if !strings.Contains(got, pattern+"\n") {
-				t.Fatalf("%s missing pattern %q; content:\n%s", gitignorePath, pattern, got)
-			}
-		}
-		if strings.Contains(got, "\n.kit/\n") || strings.HasPrefix(got, ".kit/\n") {
-			t.Fatalf("%s should not ignore all of .kit/; content:\n%s", gitignorePath, got)
-		}
-	})
-}
-
-func TestRunInit_CreatesLocalEnvironmentFiles(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		envContent, err := os.ReadFile(filepath.Join(tempDir, envPath))
-		if err != nil {
-			t.Fatalf("expected %s to be created: %v", envPath, err)
-		}
-		if string(envContent) != "" {
-			t.Fatalf("%s content = %q, want blank file", envPath, envContent)
-		}
-
-		envrcContent, err := os.ReadFile(filepath.Join(tempDir, envrcPath))
-		if err != nil {
-			t.Fatalf("expected %s to be created: %v", envrcPath, err)
-		}
-		if string(envrcContent) != templates.Envrc {
-			t.Fatalf("%s content = %q, want %q", envrcPath, envrcContent, templates.Envrc)
-		}
-	})
-}
-
-func TestRunInit_PreservesExistingLocalEnvironmentFiles(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	existingEnv := "CUSTOM=value\n"
-	existingEnvrc := "source_env .custom\n"
-	if err := os.WriteFile(filepath.Join(tempDir, envPath), []byte(existingEnv), 0644); err != nil {
-		t.Fatalf("failed to seed %s: %v", envPath, err)
+	created, err := config.Load(tempDir)
+	if err != nil {
+		t.Fatalf("config.Load() error = %v", err)
 	}
-	if err := os.WriteFile(filepath.Join(tempDir, envrcPath), []byte(existingEnvrc), 0644); err != nil {
-		t.Fatalf("failed to seed %s: %v", envrcPath, err)
-	}
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		envContent, err := os.ReadFile(filepath.Join(tempDir, envPath))
-		if err != nil {
-			t.Fatalf("failed to read %s: %v", envPath, err)
-		}
-		if string(envContent) != existingEnv {
-			t.Fatalf("%s content = %q, want %q", envPath, envContent, existingEnv)
-		}
-
-		envrcContent, err := os.ReadFile(filepath.Join(tempDir, envrcPath))
-		if err != nil {
-			t.Fatalf("failed to read %s: %v", envrcPath, err)
-		}
-		if string(envrcContent) != existingEnvrc {
-			t.Fatalf("%s content = %q, want %q", envrcPath, envrcContent, existingEnvrc)
-		}
-	})
-}
-
-func TestRunInit_PreservesExistingCodeRabbitConfig(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	existing := "reviews:\n  auto_review:\n    enabled: false\n"
-	if err := os.WriteFile(filepath.Join(tempDir, codeRabbitConfigPath), []byte(existing), 0644); err != nil {
-		t.Fatalf("failed to seed %s: %v", codeRabbitConfigPath, err)
-	}
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		content, err := os.ReadFile(filepath.Join(tempDir, codeRabbitConfigPath))
-		if err != nil {
-			t.Fatalf("failed to read %s: %v", codeRabbitConfigPath, err)
-		}
-		if string(content) != existing {
-			t.Fatalf("%s content = %q, want %q", codeRabbitConfigPath, content, existing)
-		}
-	})
-}
-
-func TestRunInit_AppendsMissingGitignoreEntries(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	existing := "# Custom ignores\ncustom.log\n.kit/runs/\n"
-	if err := os.WriteFile(filepath.Join(tempDir, gitignorePath), []byte(existing), 0644); err != nil {
-		t.Fatalf("failed to seed %s: %v", gitignorePath, err)
-	}
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		content, err := os.ReadFile(filepath.Join(tempDir, gitignorePath))
-		if err != nil {
-			t.Fatalf("failed to read %s: %v", gitignorePath, err)
-		}
-		got := string(content)
-		if !strings.HasPrefix(got, existing) {
-			t.Fatalf("expected existing content to be preserved, got:\n%s", got)
-		}
-		for _, pattern := range kitGitignorePatterns() {
-			if !strings.Contains(got, pattern+"\n") {
-				t.Fatalf("%s missing pattern %q; content:\n%s", gitignorePath, pattern, got)
-			}
-		}
-		if strings.Count(got, ".kit/runs/") != 1 {
-			t.Fatalf("expected .kit/runs/ to remain deduplicated, got:\n%s", got)
-		}
-	})
-}
-
-func TestRunInit_PreservesExistingPullRequestTemplate(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	existing := "## Summary\n\nCustom template\n"
-	if err := document.Write(filepath.Join(tempDir, pullRequestTemplatePath), existing); err != nil {
-		t.Fatalf("failed to seed %s: %v", pullRequestTemplatePath, err)
-	}
-
-	withInitFlags(t, func() {
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-
-		content, err := os.ReadFile(filepath.Join(tempDir, pullRequestTemplatePath))
-		if err != nil {
-			t.Fatalf("failed to read %s: %v", pullRequestTemplatePath, err)
-		}
-		if string(content) != existing {
-			t.Fatalf("%s content = %q, want %q", pullRequestTemplatePath, content, existing)
-		}
-	})
+	assertDefaultInitLoopAgent(t, created)
 }
 
 func TestRunInitRefresh_MigratesGeneratedVerboseInstructionsAndCreatesManagedFiles(t *testing.T) {
@@ -493,578 +267,6 @@ func TestRunInitRefresh_MigratesGeneratedVerboseInstructionsAndCreatesManagedFil
 	}
 }
 
-func TestRunInitRefresh_FileForceOverwritesOnlySelectedExistingFile(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	writeFile(t, filepath.Join(tempDir, envrcPath), "source_env .custom\n")
-	writeFile(t, filepath.Join(tempDir, codeRabbitConfigPath), "custom coderabbit\n")
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initForce = true
-		initOutputOnly = true
-		initRefreshFiles = []string{envrcPath}
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	envrcContent, err := os.ReadFile(filepath.Join(tempDir, envrcPath))
-	if err != nil {
-		t.Fatalf("failed to read %s: %v", envrcPath, err)
-	}
-	if string(envrcContent) != templates.Envrc {
-		t.Fatalf("%s content = %q, want %q", envrcPath, envrcContent, templates.Envrc)
-	}
-
-	codeRabbitContent, err := os.ReadFile(filepath.Join(tempDir, codeRabbitConfigPath))
-	if err != nil {
-		t.Fatalf("failed to read %s: %v", codeRabbitConfigPath, err)
-	}
-	if string(codeRabbitContent) != "custom coderabbit\n" {
-		t.Fatalf("%s content = %q, want custom content", codeRabbitConfigPath, codeRabbitContent)
-	}
-	assertFileDoesNotExist(t, filepath.Join(tempDir, agentsMDPath))
-}
-
-func TestRunInitRefresh_ForceDoesNotOverwriteExistingScaffoldFilesWithoutFileTarget(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	stubRulesetRegistry(t)
-
-	writeFile(t, filepath.Join(tempDir, envrcPath), "source_env .custom\n")
-	writeFile(t, filepath.Join(tempDir, "docs", "agents", "GUARDRAILS.md"), "# Guardrails\n\nold\n")
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initForce = true
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	envrcContent, err := os.ReadFile(filepath.Join(tempDir, envrcPath))
-	if err != nil {
-		t.Fatalf("failed to read %s: %v", envrcPath, err)
-	}
-	if string(envrcContent) != "source_env .custom\n" {
-		t.Fatalf("%s content = %q, want custom content", envrcPath, envrcContent)
-	}
-
-	guardrailsContent, err := os.ReadFile(filepath.Join(tempDir, "docs", "agents", "GUARDRAILS.md"))
-	if err != nil {
-		t.Fatalf("failed to read GUARDRAILS.md: %v", err)
-	}
-	if string(guardrailsContent) != initTestSupportFileContent("docs/agents/GUARDRAILS.md") {
-		t.Fatalf("expected generated docs support file to be overwritten on force, got:\n%s", guardrailsContent)
-	}
-}
-
-func TestRunInitRefresh_RejectsUnsupportedFileTarget(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initRefreshFiles = []string{"README.md"}
-
-		err := runInit(initCmd, nil)
-		if err == nil || !strings.Contains(err.Error(), "not a Kit-managed refresh target") {
-			t.Fatalf("expected unsupported target error, got %v", err)
-		}
-	})
-}
-
-func TestRunInitRefresh_DryRunDiffPrintsPlannedChangesWithoutWriting(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-
-	existing := "source_env .custom\n"
-	writeFile(t, filepath.Join(tempDir, envrcPath), existing)
-
-	var output string
-	withInitFlags(t, func() {
-		initRefresh = true
-		initDryRun = true
-		initDiff = true
-		initForce = true
-		initRefreshFiles = []string{envrcPath}
-
-		output = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	for _, check := range []string{
-		"diff --git a/.envrc b/.envrc",
-		"--- a/.envrc",
-		"+++ b/.envrc",
-		"-source_env .custom",
-		"+dotenv_if_exists",
-		"Dry run complete. Planned Created: 0, Updated: 1, Merged: 0, Skipped: 0",
-	} {
-		if !strings.Contains(output, check) {
-			t.Fatalf("expected dry-run diff output to contain %q, got:\n%s", check, output)
-		}
-	}
-
-	content, err := os.ReadFile(filepath.Join(tempDir, envrcPath))
-	if err != nil {
-		t.Fatalf("failed to read %s: %v", envrcPath, err)
-	}
-	if string(content) != existing {
-		t.Fatalf("dry run wrote %s; content = %q, want %q", envrcPath, content, existing)
-	}
-}
-
-func TestRunInitRefresh_DryRunDoesNotWritePlannedRefresh(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	stubRulesetRegistry(t)
-
-	cfg := config.Default()
-	cfg.InstructionScaffoldVersion = config.InstructionScaffoldVersionVerbose
-	if err := config.Save(tempDir, cfg); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	writeFile(t, filepath.Join(tempDir, agentsMDPath), templates.LegacyAgentsMD)
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initDryRun = true
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	unchanged, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	if unchanged.InstructionScaffoldVersion != config.InstructionScaffoldVersionVerbose {
-		t.Fatalf("dry run updated config version = %d", unchanged.InstructionScaffoldVersion)
-	}
-
-	agentsContent, err := os.ReadFile(filepath.Join(tempDir, agentsMDPath))
-	if err != nil {
-		t.Fatalf("failed to read %s: %v", agentsMDPath, err)
-	}
-	if string(agentsContent) != templates.LegacyAgentsMD {
-		t.Fatalf("dry run updated %s", agentsMDPath)
-	}
-	assertFileDoesNotExist(t, filepath.Join(tempDir, "docs", "agents", "README.md"))
-}
-
-func TestRunInitRefresh_AdoptsExistingStatusOnlyRulesetAsManaged(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	registry := registryRulesetForTest("safety-guardrails", []string{"git", "github"})
-	stubRulesetRegistry(t, registry)
-
-	if err := config.Save(tempDir, config.Default()); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	local := strings.Replace(registry.Content, "status: active", "status: optional", 1)
-	writeFile(t, filepath.Join(tempDir, rulesetTarget(registry.Slug)), local)
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initOutputOnly = true
-		initRefreshFiles = []string{rulesetTarget(registry.Slug)}
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	content, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget(registry.Slug)))
-	if err != nil {
-		t.Fatalf("failed to read ruleset: %v", err)
-	}
-	if !strings.Contains(string(content), "status: optional") {
-		t.Fatalf("expected local status to remain optional, got:\n%s", content)
-	}
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	artifact, ok := updated.RegistryArtifact(rulesetKind, registry.Slug)
-	if !ok {
-		t.Fatalf("expected registry artifact for %s", registry.Slug)
-	}
-	if artifact.State != registryArtifactStateManaged || artifact.InstalledHash != registry.NormalizedHash {
-		t.Fatalf("artifact = %#v, want managed hash %s", artifact, registry.NormalizedHash)
-	}
-}
-
-func TestRunInitRefresh_InstallsDownstreamCapabilitiesUsageRuleNotMaintainerRule(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	usage := registryRulesetWithContentForTest("kit-capabilities-usage", downstreamCapabilitiesUsageRulesetForTest(), "test-usage-commit")
-	maintainer := registryRulesetForTest("command-capabilities", []string{"kit", "cli", "capabilities"})
-	maintainer.Metadata.RegistryScope = rulesetRegistryScopeKitMaintainer
-	maintainer.Content = strings.Replace(
-		maintainer.Content,
-		"## Rules",
-		"## Rules\n\n- Update `pkg/cli/capabilities_catalog.go`.",
-		1,
-	)
-	stubRulesetRegistry(t, usage, maintainer)
-
-	if err := config.Save(tempDir, config.Default()); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initOutputOnly = true
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	usageContent, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget("kit-capabilities-usage")))
-	if err != nil {
-		t.Fatalf("expected downstream usage ruleset to be installed: %v", err)
-	}
-	for _, check := range []string{
-		"slug: kit-capabilities-usage",
-		"kit capabilities <command> --json",
-		"Do not maintain Kit's internal command catalog from a downstream project",
-	} {
-		if !strings.Contains(string(usageContent), check) {
-			t.Fatalf("expected downstream usage ruleset to contain %q, got:\n%s", check, usageContent)
-		}
-	}
-	if document.Exists(filepath.Join(tempDir, rulesetTarget("command-capabilities"))) {
-		t.Fatalf("maintainer-only command-capabilities ruleset should not be installed in downstream refresh")
-	}
-
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	if _, ok := updated.RegistryArtifact(rulesetKind, "kit-capabilities-usage"); !ok {
-		t.Fatalf("expected registry state for downstream usage ruleset")
-	}
-	if _, ok := updated.RegistryArtifact(rulesetKind, "command-capabilities"); ok {
-		t.Fatalf("did not expect registry state for maintainer-only ruleset")
-	}
-}
-
-func TestRunInitRefresh_DryRunDiffReportsDownstreamCapabilitiesUsageRuleAdoption(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	usage := registryRulesetWithContentForTest("kit-capabilities-usage", downstreamCapabilitiesUsageRulesetForTest(), "test-usage-commit")
-	stubRulesetRegistry(t, usage)
-
-	if err := config.Save(tempDir, config.Default()); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-
-	var output string
-	withInitFlags(t, func() {
-		initRefresh = true
-		initDryRun = true
-		initDiff = true
-		initRefreshFiles = []string{rulesetTarget("kit-capabilities-usage")}
-
-		output = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	for _, check := range []string{
-		"diff --git a/docs/references/rules/kit-capabilities-usage.md b/docs/references/rules/kit-capabilities-usage.md",
-		"+slug: kit-capabilities-usage",
-		"+- Use `kit capabilities` for command discovery.",
-		"Dry run complete. Planned Created:",
-	} {
-		if !strings.Contains(output, check) {
-			t.Fatalf("expected dry-run diff output to contain %q, got:\n%s", check, output)
-		}
-	}
-	assertFileDoesNotExist(t, filepath.Join(tempDir, rulesetTarget("kit-capabilities-usage")))
-}
-
-func TestRunInitRefresh_AdoptsExistingCustomRulesetWithoutOverwriting(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	registry := registryRulesetForTest("work-lane-gating", []string{"workflow"})
-	stubRulesetRegistry(t, registry)
-
-	if err := config.Save(tempDir, config.Default()); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	local := strings.Replace(registry.Content, "## Examples", "Local custom guidance.\n\n## Examples", 1)
-	writeFile(t, filepath.Join(tempDir, rulesetTarget(registry.Slug)), local)
-
-	var output string
-	withInitFlags(t, func() {
-		initRefresh = true
-		initRefreshFiles = []string{rulesetTarget(registry.Slug)}
-
-		output = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	content, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget(registry.Slug)))
-	if err != nil {
-		t.Fatalf("failed to read ruleset: %v", err)
-	}
-	if string(content) != local {
-		t.Fatalf("custom ruleset was overwritten:\n%s", content)
-	}
-	if !strings.Contains(output, "local custom content") {
-		t.Fatalf("expected local-custom note, got:\n%s", output)
-	}
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	artifact, ok := updated.RegistryArtifact(rulesetKind, registry.Slug)
-	if !ok || artifact.State != registryArtifactStateLocalCustom {
-		t.Fatalf("artifact = %#v, want local-custom", artifact)
-	}
-}
-
-func TestRunInitRefresh_ForceAcceptsRegistryContentAndPreservesStatus(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	registry := registryRulesetForTest("github-pr-delivery", []string{"github"})
-	stubRulesetRegistry(t, registry)
-
-	if err := config.Save(tempDir, config.Default()); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	local := strings.Replace(registry.Content, "status: active", "status: optional", 1)
-	local = strings.Replace(local, "## Examples", "Local custom guidance.\n\n## Examples", 1)
-	writeFile(t, filepath.Join(tempDir, rulesetTarget(registry.Slug)), local)
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initForce = true
-		initOutputOnly = true
-		initRefreshFiles = []string{rulesetTarget(registry.Slug)}
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	content, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget(registry.Slug)))
-	if err != nil {
-		t.Fatalf("failed to read ruleset: %v", err)
-	}
-	if strings.Contains(string(content), "Local custom guidance.") {
-		t.Fatalf("expected force to replace custom content, got:\n%s", content)
-	}
-	if !strings.Contains(string(content), "status: optional") {
-		t.Fatalf("expected force to preserve local status, got:\n%s", content)
-	}
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	artifact, ok := updated.RegistryArtifact(rulesetKind, registry.Slug)
-	if !ok || artifact.State != registryArtifactStateManaged || artifact.InstalledHash != registry.NormalizedHash {
-		t.Fatalf("artifact = %#v, want managed hash %s", artifact, registry.NormalizedHash)
-	}
-}
-
-func TestRunInitRefresh_FastForwardsManagedRuleset(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	base := registryRulesetForTest("safety-guardrails", []string{"git", "github"})
-	remoteContent := strings.Replace(base.Content, "## Verification", "- Registry refresh verification.\n\n## Verification", 1)
-	remote := registryRulesetWithContentForTest(base.Slug, remoteContent, "new-commit")
-	stubRulesetRegistry(t, remote)
-
-	cfg := config.Default()
-	recordRulesetRegistryState(cfg, base, registryArtifactStateManaged, base.NormalizedHash, base.Content)
-	if err := config.Save(tempDir, cfg); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	writeFile(t, filepath.Join(tempDir, rulesetTarget(base.Slug)), base.Content)
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initOutputOnly = true
-		initRefreshFiles = []string{rulesetTarget(base.Slug)}
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	content, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget(base.Slug)))
-	if err != nil {
-		t.Fatalf("failed to read ruleset: %v", err)
-	}
-	if !strings.Contains(string(content), "Registry refresh verification.") {
-		t.Fatalf("expected remote content to be applied, got:\n%s", content)
-	}
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	artifact, ok := updated.RegistryArtifact(rulesetKind, base.Slug)
-	if !ok || artifact.InstalledHash != remote.NormalizedHash {
-		t.Fatalf("artifact = %#v, want hash %s", artifact, remote.NormalizedHash)
-	}
-}
-
-func TestRunInitRefresh_SectionMergesManagedRuleset(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	base := registryRulesetForTest("work-lane-gating", []string{"workflow"})
-	localContent := strings.Replace(base.Content, "## Examples", "- Local example.\n\n## Examples", 1)
-	remoteContent := strings.Replace(base.Content, "## Verification", "- Remote verification.\n\n## Verification", 1)
-	remote := registryRulesetWithContentForTest(base.Slug, remoteContent, "new-commit")
-	stubRulesetRegistry(t, remote)
-	stubRulesetRegistryContent(t, map[string]string{base.SourceCommit: base.Content})
-
-	cfg := config.Default()
-	recordRulesetRegistryState(cfg, base, registryArtifactStateManaged, base.NormalizedHash, base.Content)
-	if err := config.Save(tempDir, cfg); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	writeFile(t, filepath.Join(tempDir, rulesetTarget(base.Slug)), localContent)
-
-	withInitFlags(t, func() {
-		initRefresh = true
-		initOutputOnly = true
-		initRefreshFiles = []string{rulesetTarget(base.Slug)}
-
-		_ = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	content, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget(base.Slug)))
-	if err != nil {
-		t.Fatalf("failed to read ruleset: %v", err)
-	}
-	for _, check := range []string{"Local example.", "Remote verification."} {
-		if !strings.Contains(string(content), check) {
-			t.Fatalf("expected merged content to contain %q, got:\n%s", check, content)
-		}
-	}
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	artifact, ok := updated.RegistryArtifact(rulesetKind, base.Slug)
-	if !ok || artifact.State != registryArtifactStateManaged {
-		t.Fatalf("artifact = %#v, want managed", artifact)
-	}
-	mergedHash, err := normalizedRulesetContentHash(string(content), remote.Metadata.Status)
-	if err != nil {
-		t.Fatalf("merged hash error: %v", err)
-	}
-	if artifact.InstalledHash != mergedHash {
-		t.Fatalf("artifact.InstalledHash = %s, want merged hash %s", artifact.InstalledHash, mergedHash)
-	}
-	if artifact.InstalledHash == remote.NormalizedHash {
-		t.Fatalf("section-merged artifact stored remote hash instead of merged content hash")
-	}
-	if len(artifact.Sections) == 0 {
-		t.Fatalf("section-merged artifact should store section hashes for future refreshes")
-	}
-}
-
-func TestRunInitRefresh_SkipsConflictedManagedRuleset(t *testing.T) {
-	tempDir := t.TempDir()
-	setupInitHome(t)
-	setWorkingDirectory(t, tempDir)
-	base := registryRulesetForTest("github-pr-delivery", []string{"github"})
-	localContent := strings.Replace(base.Content, "## Rules", "- Local rule change.\n\n## Rules", 1)
-	remoteContent := strings.Replace(base.Content, "## Rules", "- Remote rule change.\n\n## Rules", 1)
-	remote := registryRulesetWithContentForTest(base.Slug, remoteContent, "new-commit")
-	stubRulesetRegistry(t, remote)
-	stubRulesetRegistryContent(t, map[string]string{base.SourceCommit: base.Content})
-
-	cfg := config.Default()
-	recordRulesetRegistryState(cfg, base, registryArtifactStateManaged, base.NormalizedHash, base.Content)
-	if err := config.Save(tempDir, cfg); err != nil {
-		t.Fatalf("failed to save config: %v", err)
-	}
-	writeFile(t, filepath.Join(tempDir, rulesetTarget(base.Slug)), localContent)
-
-	var output string
-	withInitFlags(t, func() {
-		initRefresh = true
-		initRefreshFiles = []string{rulesetTarget(base.Slug)}
-
-		output = captureStdout(t, func() {
-			if err := runInit(initCmd, nil); err != nil {
-				t.Fatalf("runInit() error = %v", err)
-			}
-		})
-	})
-
-	content, err := os.ReadFile(filepath.Join(tempDir, rulesetTarget(base.Slug)))
-	if err != nil {
-		t.Fatalf("failed to read ruleset: %v", err)
-	}
-	if string(content) != localContent {
-		t.Fatalf("conflicted ruleset was overwritten:\n%s", content)
-	}
-	if !strings.Contains(output, "changed locally and in registry") {
-		t.Fatalf("expected conflict note, got:\n%s", output)
-	}
-	updated, err := config.Load(tempDir)
-	if err != nil {
-		t.Fatalf("config.Load() error = %v", err)
-	}
-	artifact, ok := updated.RegistryArtifact(rulesetKind, base.Slug)
-	if !ok || artifact.State != registryArtifactStateConflict {
-		t.Fatalf("artifact = %#v, want conflict", artifact)
-	}
-}
-
 func TestRunInit_DiffRequiresDryRun(t *testing.T) {
 	tempDir := t.TempDir()
 	setupInitHome(t)
@@ -1079,99 +281,4 @@ func TestRunInit_DiffRequiresDryRun(t *testing.T) {
 			t.Fatalf("expected --diff without --dry-run error, got %v", err)
 		}
 	})
-}
-
-func initTestSupportFileContent(relativePath string) string {
-	for _, file := range templates.InstructionSupportFiles(config.InstructionScaffoldVersionTOC) {
-		if file.RelativePath == relativePath {
-			return file.Content
-		}
-	}
-	return ""
-}
-
-func withInitFlags(t *testing.T, run func()) {
-	t.Helper()
-
-	originalCopy := initCopy
-	originalOutputOnly := initOutputOnly
-	originalRefresh := initRefresh
-	originalForce := initForce
-	originalDryRun := initDryRun
-	originalDiff := initDiff
-	originalRefreshFiles := initRefreshFiles
-
-	t.Cleanup(func() {
-		initCopy = originalCopy
-		initOutputOnly = originalOutputOnly
-		initRefresh = originalRefresh
-		initForce = originalForce
-		initDryRun = originalDryRun
-		initDiff = originalDiff
-		initRefreshFiles = originalRefreshFiles
-	})
-
-	initCopy = false
-	initOutputOnly = false
-	initRefresh = false
-	initForce = false
-	initDryRun = false
-	initDiff = false
-	initRefreshFiles = nil
-
-	run()
-}
-
-func setupInitHome(t *testing.T) string {
-	t.Helper()
-
-	homeDir := t.TempDir()
-	t.Setenv("HOME", homeDir)
-	return homeDir
-}
-
-func downstreamCapabilitiesUsageRulesetForTest() string {
-	return `---
-kind: ruleset
-slug: kit-capabilities-usage
-description: Downstream Kit command discovery guidance.
-status: active
-registry_scope: downstream
-applies_to:
-  - kit
-  - cli
-  - command-discovery
-read_policy_default: conditional
----
-
-# Ruleset: kit-capabilities-usage
-
-## Purpose
-
-- Use ` + "`kit capabilities`" + ` for downstream command discovery.
-
-## Applies When
-
-- A downstream project needs to choose a Kit command.
-
-## Rules
-
-- Use ` + "`kit capabilities`" + ` for command discovery.
-- Prefer ` + "`kit capabilities <command> --json`" + ` after narrowing the command.
-- Do not maintain Kit's internal command catalog from a downstream project.
-
-## Anti-Patterns
-
-- Do not tell downstream projects to edit ` + "`pkg/cli/capabilities_catalog.go`" + `.
-
-## Verification
-
-- ` + "`kit capabilities <command> --json`" + ` describes the selected command.
-
-## Examples
-
-` + "```bash" + `
-kit capabilities dispatch --json
-` + "```" + `
-`
 }
